+++ src/extension.ts
@@ -1,36 +1,39 @@
 'use strict';
 
 import * as vscode from 'vscode';
-import { luoguManager } from './luoguManager';
+import { luoguUserManager } from './luoguUserManager';
 import { luoguStatusBarItem } from './luoguStatusBarItem';
-import { searchProblem } from './commands/search';
-import { submitSolution } from './commands/submit';
-import { login } from './commands/login';
+import { search } from './commands/search';
+import { submit } from './commands/submit';
+import { promptForOpenOutputChannel, DialogType } from './utils/uiUtils';
 
 export function activate(context: vscode.ExtensionContext) {
     const channel: vscode.OutputChannel = vscode.window.createOutputChannel("Luogu");
 
     let about = function () {
-        vscode.window.showInformationMessage('Developed by Himself65');
+        promptForOpenOutputChannel('Developed by Himself65', DialogType.info, channel);
     };
 
     let notice = function () {
-        vscode.window.showInformationMessage('还没有写这部分，敬请期待');
+        promptForOpenOutputChannel('这部分还没有写完QAQ', DialogType.info, channel);
     };
 
+    // TAG: showProblem 与 searchProblem 功能不同
+    // 前者通过 Explorer 访问题目，后者直接搜索题目
+    // TODO: 之后会整合功能
     context.subscriptions.push(
         // vscode.window.registerTreeDataProvider("luoguExplorer", luoguTreeDataProvider),
         vscode.commands.registerCommand("luogu.about", about),
-        vscode.commands.registerCommand("luogu.signin", () => login(channel)),
-        vscode.commands.registerCommand("luogu.signout", notice),
+        vscode.commands.registerCommand("luogu.signin", () => luoguUserManager.signIn(channel)),
+        vscode.commands.registerCommand("luogu.signout", () => luoguUserManager.signOut(channel)),
         vscode.commands.registerCommand("luogu.showProblem", notice),
-        vscode.commands.registerCommand("luogu.searchProblem", () => searchProblem(channel)),
-        vscode.commands.registerCommand("luogu.submitSolution", () => submitSolution(channel)),
+        vscode.commands.registerCommand("luogu.searchProblem", () => search(channel)),
+        vscode.commands.registerCommand("luogu.submitSolution", () => submit(channel)),
         vscode.commands.registerCommand("luogu.refreshExplorer", notice)
     );
 
-    luoguManager.on('statusChanged', () => {
-        luoguStatusBarItem.updateStatusBar(luoguManager.getStatus(), luoguManager.getUser());
+    luoguUserManager.on('statusChanged', () => {
+        luoguStatusBarItem.updateStatusBar(luoguUserManager.getStatus(), luoguUserManager.getUser());
     });
 }