+++ src/shared.ts
@@ -1,8 +1,5 @@
 "use strict";
 
-import * as vscode from 'vscode';
-import * as ClientOAuth2 from 'client-oauth2';
-
 export enum UserStatus {
     SignedIn = 1,
     SignedOut = 2,
@@ -10,38 +7,44 @@ export enum UserStatus {
 
 export const Languages = [
     "Auto",
-    "C++11",
-    "C++",
-    "C",
     "Pascal",
-    "Node.js",
-    "Java",
+    "C",
+    "C++",
+    "C++11",
     "Python2",
     "Python3",
+    "Java",
+    "Node.js"
 ];
 
 export enum ProblemState {
-    AC = 1,
-    UAC = 2,
-    Unknow = 3,
+    "Waiting" = 0,
+    "Judging" = 1,
+    "Compile Error" = 2,
+    "OLE" = 3,
+    "MLE" = 4,
+    "TLE" = 5,
+    "WA" = 6,
+    "RE" = 7,
+    "Accepted" = 12,
+    "Unaccepted" = 14,
+    "Hack Success" = 21,
+    "Hack Failure" = 22,
+    "Hack Skipped" = 23
 }
 
-export function showError(message?: string): void {
-    vscode.window.showErrorMessage(message);
+export interface OAuth2ResponseData {
+    access_token: string;
+    token_type: string;
+    expires_in: number;
+    refresh_token?: string;
+    scope?: Array<string>;
 }
 
 export const OAuthInfo = {
+    grant_type: 'password',
     clientID: 'luogu-vscode',
-    key: 'HasHidden',
-    accessTokenUri: 'https://luogu.org/api/OAuth2/accessToken',
-    authorizationUri: 'https://luogu.org/api/OAuth2/authorize',
-    scopes: []
+    client_secret: 'HasHidden',
+    accessTokenUri: 'https://www.luogu.org/api/OAuth2/accessToken',
+    authorizationUri: 'https://www.luogu.org/api/OAuth2/authorize'
 };
-
-export const luoguAuth = new ClientOAuth2({
-    clientId: OAuthInfo.clientID,
-    clientSecret: OAuthInfo.key,
-    accessTokenUri: OAuthInfo.accessTokenUri,
-    authorizationUri: OAuthInfo.authorizationUri,
-    scopes: OAuthInfo.scopes
-});